+++ src/browser/Terminal.test.ts
@@ -9,7 +9,7 @@ import { assert } from 'chai';
 import { DEFAULT_ATTR_DATA } from '../common/buffer/BufferLine';
 import { CellData } from '../common/buffer/CellData';
 import { MockUnicodeService, createCellData } from '../common/TestUtils.test';
-import { IMarker } from '../common/Types';
+import { IMarker } from '../common/buffer/Types';
 
 const INIT_COLS = 80;
 const INIT_ROWS = 24;
