+++ src/common/CoreTerminal.ts
@@ -26,7 +26,8 @@ import { InstantiationService } from './services/InstantiationService';
 import { LogService } from './services/LogService';
 import { BufferService, BufferServiceConstants } from './services/BufferService';
 import { OptionsService } from './services/OptionsService';
-import { IDisposable, IAttributeData, IScrollEvent, ITerminalOptions, IParams } from './Types';
+import { IDisposable, IScrollEvent, ITerminalOptions, IParams } from './Types';
+import { IAttributeData, IBufferSet } from './buffer/Types';
 import { CoreService } from './services/CoreService';
 import { MouseStateService } from './services/MouseStateService';
 import { UnicodeV6 } from './input/UnicodeV6';
@@ -34,7 +35,6 @@ import { UnicodeService } from './services/UnicodeService';
 import { CharsetService } from './services/CharsetService';
 import { updateWindowsModeWrappedState } from './WindowsMode';
 import { IFunctionIdentifier } from './parser/Types';
-import { IBufferSet } from './buffer/Types';
 import { InputHandler } from './InputHandler';
 import { WriteBuffer } from './input/WriteBuffer';
 import { OscLinkService } from './services/OscLinkService';
