+++ src/common/buffer/Marker.ts
@@ -3,9 +3,9 @@
  * @license MIT
  */
 
-import { IDisposable, IMarker } from '../Types';
+import { dispose, IDisposable } from '../Lifecycle';
+import { IMarker } from './Types';
 import { Emitter } from '../Event';
-import { dispose } from '../Lifecycle';
 
 export class Marker implements IMarker {
   private static _nextId = 1;
