+++ src/headless/Terminal.ts
@@ -22,9 +22,9 @@
  */
 
 import { DEFAULT_ATTR_DATA } from '../common/buffer/BufferLine';
-import { IBuffer } from '../common/buffer/Types';
+import { IBuffer, IMarker } from '../common/buffer/Types';
 import { CoreTerminal } from '../common/CoreTerminal';
-import { IMarker, ITerminalOptions } from '../common/Types';
+import { ITerminalOptions } from '../common/Types';
 import { Emitter, EventUtils } from '../common/Event';
 
 export class Terminal extends CoreTerminal {
